Starting migration to AWS Lambda
